refactor: move fragments into respective packages
