+++ code/app/src/main/java/com/hamidat/nullpointersapp/authFragments/SignUpFragment.java
@@ -1,4 +1,4 @@
-package com.hamidat.nullpointersapp.fragments;
+package com.hamidat.nullpointersapp.authFragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
