--- code/app/src/main/java/com/hamidat/nullpointersapp/fragments/HomeFeedFragment.java
@@ -1,4 +1,0 @@
-package com.hamidat.nullpointersapp.fragments;
-
-public class HomeFeedFragment {
-}
