--- code/app/src/main/java/com/hamidat/nullpointersapp/fragments/MapFragment.java
@@ -1,4 +1,0 @@
-package com.hamidat.nullpointersapp.fragments;
-
-public class MapFragment {
-}
