+++ code/app/src/main/java/com/hamidat/nullpointersapp/mainFragments/HomeFeedFragment.java
@@ -1,0 +1,4 @@
+package com.hamidat.nullpointersapp.mainFragments;
+
+public class HomeFeedFragment {
+}
