+++ code/app/src/main/java/com/hamidat/nullpointersapp/mainFragments/MapFragment.java
@@ -1,0 +1,4 @@
+package com.hamidat.nullpointersapp.mainFragments;
+
+public class MapFragment {
+}
