+++ code/app/src/main/java/com/hamidat/nullpointersapp/mainFragments/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.hamidat.nullpointersapp.fragments;
+package com.hamidat.nullpointersapp.mainFragments;
 
 import static com.hamidat.nullpointersapp.utils.AppConstants.*;
